Refac: move metrics
